Concurrencia en proceso (nunca mejor dicho) y ya inserta guachi

+++ src/main/kotlin/controllers/UserController.kt
@@ -10,67 +10,87 @@ import java.util.*
 object UserController {
     private val service = UserService()
 
-    fun findAllUsers(): String {
-        val result = GsonBuilder().setPrettyPrinting().create()
-            .toJson(service.getAllUsers())
-            ?: "Error at UserController.findAllUsers"
-        return generateRespuesta(result, "Error at UserController.findAllUsers")
+    suspend fun findAllUsers(): String {
+        //return generateRespuesta(service.getAllUsers())
+        return service.getAllUsers().toString()
     }
 
-    fun findAllUsersWithRole(role: Profile): String {
+    suspend fun findAllUsersWithRole(role: Profile): String {
+        return service.getAllUsers().filter { it.perfil == role }.toString()
+        /*
         val result = GsonBuilder().setPrettyPrinting().create()
             .toJson(service.getAllUsers().filter { it.perfil == role })
             ?: "Error at UserController.findAllUsersWithRole $role"
         return generateRespuesta(result, "Error at UserController.findAllUsersWithRole $role")
+
+         */
+
+        //return generateRespuesta(service.getAllUsers().filter { it.perfil == role })
     }
 
-    fun getUserById(id: UUID): String {
+    suspend fun getUserById(id: UUID): String {
+        return service.getUserById(id)?.toJSON() ?: "User with id $id not found."
+        /*
         val result = GsonBuilder().setPrettyPrinting().create()
             .toJson(service.getUserById(id))
             ?: "User with id $id not found."
         return generateRespuesta(result, "User with id $id not found.")
+
+         */
+
+        //return generateRespuesta(service.getUserById(id))
     }
 
-    fun getUserByEmail(email: String): String {
+    suspend fun getUserByEmail(email: String): String {
+        return service.getUserByMail(email)?.toJSON() ?: "User with email $email not found."
+        /*
         val result = GsonBuilder().setPrettyPrinting().create()
             .toJson(service.getUserByMail(email))
             ?: "User with email $email not found."
         return generateRespuesta(result, "User with email $email not found.")
+
+         */
+
+        //return generateRespuesta(service.getUserByMail(email))
     }
 
-    fun getUserByEmailForLogin(email: String): UserDTO? {
+    suspend fun getUserByEmailForLogin(email: String): UserDTO? {
         return service.getUserByMail(email)
     }
 
-    fun getUserByPhone(phone: String): String {
+    suspend fun getUserByPhone(phone: String): String {
+        return service.getUserByPhone(phone)?.toJSON() ?: "User with phone $phone not found."
+        /*
         val result = GsonBuilder().setPrettyPrinting().create()
             .toJson(service.getUserByPhone(phone))
             ?: "User with phone $phone not found."
         return generateRespuesta(result, "User with phone $phone not found.")
+
+         */
+
+        //return generateRespuesta(service.getUserByPhone(phone))
     }
 
-    fun getUserByPhoneForLogin(phone: String): UserDTO? {
+    suspend fun getUserByPhoneForLogin(phone: String): UserDTO? {
         return service.getUserByPhone(phone)
     }
 
-    fun insertUser(dto: UserDTO): String {
-        val result = GsonBuilder().setPrettyPrinting().create()
-            .toJson(service.createUser(dto))
-            ?: "Could not insert User with id ${dto.id}"
-        return generateRespuesta(result, "Could not insert User with id ${dto.id}")
+    suspend fun insertUser(dto: UserDTO): String {
+        return service.createUser(dto).toJSON()
     }
 
-    fun deleteUser(dto: UserDTO): String {
+    suspend fun deleteUser(dto: UserDTO): String {
+        val result = service.deleteUser(dto)
+        return if (result) dto.toJSON()
+        else "Could not delete User with id ${dto.id}"
+        /*
         val result = GsonBuilder().setPrettyPrinting().create()
             .toJson(service.deleteUser(dto))
             ?: "Could not delete User with id ${dto.id}"
         return generateRespuesta(result, "Could not delete User with id ${dto.id}")
-    }
 
-    fun insertUserInit(dto: UserDTO): String {
-        val result = GsonBuilder().setPrettyPrinting().create()
-            .toJson(service.createUserInit(dto))
-            ?: "Could not insert User with id ${dto.id}"
-        return generateRespuesta(result, "Could not insert User with id ${dto.id}")
+         */
+
+        //return generateRespuesta(service.deleteUser(dto))
     }
 }
